update to match merge

+++ src/main/kotlin/fr/insalyon/pld/agile/model/Distance.kt
@@ -1,7 +1,7 @@
 package fr.insalyon.pld.agile.model
 
 class Distance constructor (
-       val value: Int,
+       val value: Long,
        val distanceUnit: DistanceUnit
 ){
 
@@ -15,11 +15,11 @@ class Distance constructor (
 
         val coef = (this.distanceUnit.coef /distanceUnit.coef)
 
-        return Distance((value * coef).toInt(), distanceUnit)
+        return Distance((value * coef).toLong(), distanceUnit)
     }
 
 }
 
-public val Int.km: Distance get() = Distance(this, Distance.DistanceUnit.KM)
-public val Int.dam: Distance get() = Distance(this, Distance.DistanceUnit.DAM)
-public val Int.m: Distance get() = Distance(this, Distance.DistanceUnit.M)
+public val Long.km: Distance get() = Distance(this, Distance.DistanceUnit.KM)
+public val Long.dam: Distance get() = Distance(this, Distance.DistanceUnit.DAM)
+public val Long.m: Distance get() = Distance(this, Distance.DistanceUnit.M)
